0.5.2 Update
The two remaining known issues are resolved:
* Link Battles occasionally tacking on invalid characters to names
* The lottery luck man disappearing when the spinning starts

In addition, nightly releases will now properly use a version controlled copy of the sheet, which will avoid inconsistencies.

After verifying this build is stable and tested, it may be promoted to our first 1.0.0 release, but for now we'll keep an eye on it in case it needs fixes.